+++ paramiko/kex_curve25519.py
@@ -89,8 +89,9 @@ class KexCurve25519(object):
         hm.add_mpint(K)
         H = self.hash_algo(hm.asbytes()).digest()
         self.transport._set_K_H(K, H)
-        algorithm = self.transport.host_key_type
-        sig = self.transport.get_server_key().sign_ssh_data(H, algorithm)
+        sig = self.transport.get_server_key().sign_ssh_data(
+            H, self.transport.host_key_type
+        )
         # construct reply
         m = Message()
         m.add_byte(c_MSG_KEXECDH_REPLY)